+++ src/hb-open-type.hh
@@ -702,7 +702,7 @@ struct SortedArrayOf : ArrayOf<Type, LenType>
     int min = 0, max = (int) this->len - 1;
     while (min <= max)
     {
-      int mid = (min + max) / 2;
+      int mid = ((unsigned int) min + (unsigned int) max) / 2;
       int c = arr[mid].cmp (x);
       if (c < 0)
         max = mid - 1;
@@ -825,7 +825,7 @@ struct VarSizedBinSearchArrayOf
     int min = 0, max = (int) header.nUnits - 1;
     while (min <= max)
     {
-      int mid = (min + max) / 2;
+      int mid = ((unsigned int) min + (unsigned int) max) / 2;
       const Type *p = (const Type *) (((const char *) &bytesZ) + (mid * size));
       int c = p->cmp (key);
       if (c < 0)
